separation of the movement gestion in a dedicate class and implementation of it for the enemy
